+++ changelog.md
@@ -9,7 +9,7 @@
 - Added basic pitch detection to the edit panel.
 - Use the , and . keys to change the selected slice on the edit panel.
 - For Stretch Selected Actions holding the Alt Key, and for the pitch shift buttons on the editor panel, holding the Shift+Alt keys will apply time-stretching where the pitch of the samples is preserved.
-- Repitch tool in sample editor panel.
+- Repitch tool in sample editor panel with pitched duplicate interval mode.
 
 2025.12.07 1.5.2
 - Corrected inconsistent import ordering in macOS file drops.